requirements.txt
Flask==3.0.3
python-telegram-bot==21.7
openai==1.52.0
requests==2.32.3
python-dotenv==1.0.1
redis==5.0.7  # optional for future
psycopg2-binary==2.9.10  # optional for future